+++ src/main/java/io/supertokens/inmemorydb/config/SQLiteConfig.java
@@ -164,4 +164,6 @@ public class SQLiteConfig {
     public String getDashboardSessionsTable() {
         return "dashboard_user_sessions";
     }
+
+    public String getOAuthClientTable(){ return "oauth_clients"; }
 }