+++ src/main/java/io/supertokens/oauth/exceptions/OAuthException.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2020, VRAI Labs and/or its affiliates. All rights reserved.
+ *    Copyright (c) 2024, VRAI Labs and/or its affiliates. All rights reserved.
  *
  *    This software is licensed under the Apache License, Version 2.0 (the
  *    "License") as published by the Apache Software Foundation.
@@ -14,18 +14,20 @@
  *    under the License.
  */
 
-package io.supertokens.httpRequest;
+package io.supertokens.oauth.exceptions;
 
-public class HttpResponseException extends Exception {
+import java.io.Serial;
 
-    private static final long serialVersionUID = 1L;
+public class OAuthException extends Exception{
+    @Serial
+    private static final long serialVersionUID = 1836718299845759897L;
 
-    public final int statusCode;
-    public final String rawMessage;
+    public final String error;
+    public final String errorDescription;
 
-    HttpResponseException(int statusCode, String message) {
-        super("Http error. Status Code: " + statusCode + ". Message: " + message);
-        this.statusCode = statusCode;
-        this.rawMessage = message;
+    public OAuthException(String error, String errorDescription){
+        super(error);
+        this.error = error;
+        this.errorDescription = errorDescription;
     }
 }