+++ CHANGELOG.md
@@ -7,6 +7,25 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Changes
+
+- Added new feature in license key: `OAUTH`
+- Adds new core config:
+    - `oauth_provider_public_service_url`
+    - `oauth_provider_admin_service_url`
+    - `oauth_provider_consent_login_base_url`
+    - `oauth_provider_url_configured_in_hydra`
+- Adds POST `/recipe/oauth/auth` for OAuth2 auth flow support
+- Adds POST `/recipe/oauth/clients` for OAuth2 client registration
+- Adds GET `/recipe/oauth/clients?clientId=example_id` for loading OAuth2 client
+- Adds DELETE `/recipe/oauth/clients` for deleting OAuth2 Clients
+- Creates new table `oauth_clients`
+- Introduces PATCH capabilities for core (receiving and sending PATCH requests)
+- Adds PATCH `/recipe/oauth/clients` for OAuth2 client update
+
+### Migration
+TODO: after plugin support
+
 ## [9.1.1] -2024-07-24
 
 ### Fixes
@@ -129,17 +148,19 @@ Make sure the core is already upgraded to version 8.0.0 before migrating
 If using PostgreSQL
 
 ```sql
-ALTER TABLE totp_user_devices ADD COLUMN IF NOT EXISTS created_at BIGINT default 0;
-ALTER TABLE totp_user_devices 
-  ALTER COLUMN created_at DROP DEFAULT;
+ALTER TABLE totp_user_devices
+    ADD COLUMN IF NOT EXISTS created_at BIGINT default 0;
+ALTER TABLE totp_user_devices
+    ALTER COLUMN created_at DROP DEFAULT;
 ```
 
 If using MySQL
 
 ```sql
-ALTER TABLE totp_user_devices ADD COLUMN created_at BIGINT UNSIGNED default 0;
-ALTER TABLE totp_user_devices 
-  ALTER COLUMN created_at DROP DEFAULT;
+ALTER TABLE totp_user_devices
+    ADD COLUMN created_at BIGINT UNSIGNED default 0;
+ALTER TABLE totp_user_devices
+    ALTER COLUMN created_at DROP DEFAULT;
 DROP INDEX all_auth_recipe_users_pagination_index2 ON all_auth_recipe_users;
 DROP INDEX all_auth_recipe_users_pagination_index4 ON all_auth_recipe_users;
 ```
@@ -191,8 +212,8 @@ For MySQL:
 ALTER TABLE user_roles DROP FOREIGN KEY user_roles_ibfk_1;
 ALTER TABLE user_roles DROP FOREIGN KEY user_roles_ibfk_2;
 ALTER TABLE user_roles
-  ADD FOREIGN KEY (app_id, tenant_id)
-    REFERENCES tenants (app_id, tenant_id) ON DELETE CASCADE;
+    ADD FOREIGN KEY (app_id, tenant_id)
+        REFERENCES tenants (app_id, tenant_id) ON DELETE CASCADE;
 ```
 
 ## [7.0.18] - 2024-02-19